+++ src/Console/Generators/MakeControllerCommand.php
@@ -15,6 +15,7 @@ class MakeControllerCommand extends GeneratorCommand
     protected $signature = 'make:module:controller
     	{slug : The slug of the module}
     	{name : The name of the controller class}
+    	{--api : Generate a module API resource controller class (exclude create & edit method)}
     	{--resource : Generate a module resource controller class}
     	{--location= : The modules location to create the module controller class in}';
 
@@ -39,6 +40,10 @@ class MakeControllerCommand extends GeneratorCommand
      */
     protected function getStub()
     {
+        if ($this->option('api')) {
+            return __DIR__ . '/stubs/controller.api.stub';
+        }
+
         if ($this->option('resource')) {
             return __DIR__ . '/stubs/controller.resource.stub';
         }
@@ -50,6 +55,7 @@ class MakeControllerCommand extends GeneratorCommand
      * Get the default namespace for the class.
      *
      * @param string $rootNamespace
+     *
      * @return string
      * @throws \Caffeinated\Modules\Exceptions\ModuleNotFoundException
      */