+++ tests/Console/Generators/__snapshots__/CommandMakeResourceTest__it_can_generate_a_new_resource_collection_with_custom_module_namespace__1.php
@@ -1,0 +1,20 @@
+<?php return '<?php
+
+namespace App\\CustomResourceNamespace\\Resource\\Http\\Resources;
+
+use Illuminate\\Http\\Resources\\Json\\ResourceCollection;
+
+class CustomResource extends ResourceCollection
+{
+    /**
+     * Transform the resource collection into an array.
+     *
+     * @param  \\Illuminate\\Http\\Request  $request
+     * @return array
+     */
+    public function toArray($request)
+    {
+        return parent::toArray($request);
+    }
+}
+';