+++ CHANGELOG.md
@@ -5,7 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this package adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added 
+- `make:module:resource` command
+- `--api` option for `make:module:controller`
+### Fixed
+- Replaced deprecated exception annotations with methods (RepositoryTest)
 ## [6.2.0 - 2020-04-02]
 ### Added
 - Laravel 7 support